rename layout folder
