+++ src/App.js
@@ -5,10 +5,10 @@ import {
 } from 'react-router-dom'
 
 // components
-import Banner from "js/layout/Banner"
-import I18nBanner from "js/layout/I18nBanner"
-import Header from "js/layout/Header"
-import Footer from "js/layout/Footer"
+import Banner from "js/page_sections/Banner"
+import I18nBanner from "js/page_sections/I18nBanner"
+import Header from "js/page_sections/Header"
+import Footer from "js/page_sections/Footer"
 
 // page routes
 import Home from "js/pages/Home"
