+++ src/css/coa.scss
@@ -22,8 +22,8 @@
 @import 'css/modules/HtmlFromAdmin';
 
 // layout
-@import 'css/layout/main';
-@import "css/layout/Header";
-@import "css/layout/Footer";
-@import "css/layout/I18nBanner";
-@import "css/layout/Navmenu";
+@import 'css/page_sections/main';
+@import "css/page_sections/Header";
+@import "css/page_sections/Footer";
+@import "css/page_sections/I18nBanner";
+@import "css/page_sections/Navmenu";
