+++ src/css/page_sections/_I18nBanner.scss
