+++ src/js/page_sections/Banner.js
