+++ src/js/page_sections/Header.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 import SearchSVG from 'js/svg/Search';
-import Navmenu from 'js/layout/Navmenu';
+import Navmenu from 'js/page_sections/Navmenu';
 
 class Header extends Component {
   constructor(props) {
