+++ src/js/page_sections/I18nBanner.js
